requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-auth==2.37.0
requests==2.32.3

